requirements.txt
matplotlib~=3.5.2
opencv-python~=4.5.5.64
numpy~=1.22.3
Flask~=2.1.2
tensorflow-gpu~=2.9.0
Pillow~=9.1.1